feat: simplify routing, redirect /[locale] to /[locale/home
