+++ src/app/[locale]/(main)/home/page.tsx
