+++ src/app/[locale]/(main)/page.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import { redirect, usePathname } from 'next/navigation';
+
+export default function Home() {
+  const pathname = usePathname();
+  redirect(`${pathname}/home`);
+}
