+++ src/app/[locale]/(main)/view/[viewId]/page.tsx
